workers/python/requirements-slim.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
dataclasses-json==0.5.7
Deprecated==1.2.14
frozenlist==1.4.1
googleapis-common-protos==1.63.2
idna==3.7
marshmallow==3.21.1
marshmallow-enum==1.5.1
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.5.6
opentelemetry-api==1.15.0
opentelemetry-exporter-otlp-proto-http==1.15.0
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
packaging==24.0
plotly==5.19.0
prometheus-client==0.14.1
protobuf==4.25.3
psutil==5.9.8
pydantic==1.10.14
pydash==6.0.2
python-dotenv==0.21.1
redis==4.6.0
requests==2.32.0
requests-futures==1.0.1
simplejson==3.19.2
six==1.16.0
tenacity==8.2.3
thrift==0.16.0
typing-inspect==0.9.0
typing_extensions==4.10.0
ujson==5.10.0
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4
